pyproject.toml
[tool.poetry]
name = "tradeapis"
version = "1.0.2"
description = "Stock Market Data Utilities and API Wrappers for Tradier, Polygon"
authors = ["Matt Stancliff <matt@genges.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
orjson = "^3.5.2"
pandas = "^1.2.4"
websockets = "^9.1"
arrow = "^1.1.0"
bs4 = "^0.0.1"
lark = "^0.11.3"
prettyprinter = "^0.18.0"
cachetools = "^4.2.2"
pandas-market-calendars = "^2.0"
loguru = "^0.5.3"

mutil = { git = "https://github.com/mattsta/mutil.git", tag="1.2.2" }
# mutil = { path = "../../mutil/", develop = true }
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
